Moved level.rs into levels folder
