+++ src/levels/mod.rs
@@ -1,0 +1,2 @@
+pub use self::level::{Level, MapIdx};
+mod level;
