+++ src/lib.rs
@@ -4,7 +4,7 @@
 
 pub mod input;
 pub mod entity;
-pub mod level;
+pub mod levels;
 pub mod misc;
 pub mod game;
 pub mod traits;
